% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_at_risk.R
\name{calculate_at_risk}
\alias{calculate_at_risk}
\title{At risk}
\usage{
calculate_at_risk(times, survival_times)
}
\arguments{
\item{times}{Vector of time points}

\item{survival_times}{Survival times (possibly censored)}
}
\value{
\code{at_risk} Vector of number of individuals at risk at each time in the original times.
}
\description{
Given survival times and a sequence of (possibly more granular) times, calculate how
many is at risk at any given time. An individual is at risk at time t if it has survival time >= t.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}
