% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_observations.R
\name{censor_observations}
\alias{censor_observations}
\title{Censor observations}
\usage{
censor_observations(uncensored_survival_times, censoring_times)
}
\arguments{
\item{uncensored_survival_times}{Vector of uncensored survival times}

\item{censoring_times}{Vector of times to censor}
}
\value{
\code{t} A vector of survival times, censored according to censoring times

\code{delta} Censoring observation vector
}
\description{
Censor survival times.
}
\examples{
Code here.
}
\keyword{keywords}
