% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_optimizable_function.R
\name{data_to_optimizable_function}
\alias{data_to_optimizable_function}
\title{Function taking design matrices \code{X}, \code{Z} and data \code{t}, {delta}, and returns a function to be optimized in nlm.}
\usage{
data_to_optimizable_function(X, Z, t, delta)
}
\arguments{
\item{X}{Design matrix for \code{ln y0 = beta^T * X}}

\item{Z}{Design matrix for \code{mu = gamma^T * Z}}

\item{t}{Survival times (possibly censored)}

\item{delta}{Vector of same length as \code{t}, indicating if observations are actually observed (1) or censored (0)}
}
\value{
\code{optimizable_function} Function taking parameters as a vector (containing \code{beta}, \code{gamma}), to be optimized over.
}
\description{
Function taking design matrices \code{X}, \code{Z} and data \code{t}, {delta}, and returns a function to be optimized in nlm.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}
