% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FHT_link_function.R
\name{FHT_link_function}
\alias{FHT_link_function}
\title{FHT link function}
\usage{
FHT_link_function(X, Z)
}
\arguments{
\item{X}{Design matrix for \code{ln y0 = beta^T * X}}

\item{Z}{Design matrix for \code{mu = gamma^T * Z}}
}
\value{
\code{link_function} Function taking \code{beta}, \code{gamma} and returning \code{y0}, \code{mu}.
}
\description{
Function taking design matrices \code{X}, \code{Z} and returns link function taking \code{beta}, \code{gamma}
}
\examples{
R code here showing how your function works
}
\keyword{keywords}
