% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FHT_loglikelihood_with_data.R
\name{FHT_loglikelihood_with_data}
\alias{FHT_loglikelihood_with_data}
\title{Get function for optimization}
\usage{
FHT_loglikelihood_with_data(t, delta)
}
\arguments{
\item{t}{Survival times (possibly censored)}

\item{delta}{Corresponding observation indicator (1 = observed, 0 = censored)}
}
\value{
\code{FHT_loglikelihood} A function taking parameters \code{y0}, \code{mu}
}
\description{
Given observations (\code{t}, \code{delta}), return
a function evaluating the log likelihood for the FHT model
}
\examples{
Code examples.
}
\keyword{keywords}
