% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{generate_test_data}
\alias{generate_test_data}
\title{Generate test data}
\usage{
generate_test_data()
}
\value{
\code{list} A list of lots of things

\code{list} A list of lots of things
among others this and this
}
\description{
Generate simulated test data: Survival times and observations
}
\examples{
result <- generate_test_data()
observations <- result$observations
times <- observations$survival_times
delta <- observations$delta
# make
}
\keyword{keywords}
