% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_minus_loglik_given_observations.R
\name{get_minus_loglikelihood_given_observations}
\alias{get_minus_loglikelihood_given_observations}
\title{Get function for optimization}
\usage{
get_minus_loglikelihood_given_observations(t, delta)
}
\arguments{
\item{t}{Survival times (possibly censored)}

\item{delta}{Corresponding observation indicator (1 = observed, 0 = censored)}
}
\value{
\code{minusloglik} A function taking one argument -- a vector \code{parameters} -- which is to be optimized.
}
\description{
Given observations (\code{t}, \code{delta}), return
a function to be optimized, which evaluates the loss function
(the negative log likelihood)
}
\examples{
t <- c(0.1, 1.0)
delta <- c(0, 1)
minus_loglik <- get_minus_loglik(t, delta)
initial_parameters <- c(1, 1)
optimized_parameters <- nlm(minus_loglik, initial_parameters)$estimate
}
\keyword{keywords}
