% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier.R
\name{kaplan_meier}
\alias{kaplan_meier}
\title{Kaplan-Meier}
\usage{
kaplan_meier(times, delta)
}
\arguments{
\item{times}{Survival times (possibly censored)}

\item{delta}{Corresponding observation vector (0 is censored, 1 is actually observed)}
}
\value{
\code{kaplan_meier} Vector of Kaplan-Meier estimates
}
\description{
Non-parametric (Kaplan-Meier) estimates of the survival curve
}
\examples{
R code here showing how your function works
}
\keyword{keywords}
