% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_function_derivative_y0.R
\name{loss_function_derivative_y0}
\alias{loss_function_derivative_y0}
\title{Loss function \code{rho} differentiated wrt. \code{y0}}
\usage{
loss_function_derivative_y0(y0, mu, sigma2, t, delta)
}
\arguments{
\item{y0}{Initial level of underlying Wiener process \code{y0}}

\item{mu}{Drift of underlying Wiener process \code{mu}}
}
\value{
gradient The gradient
}
\description{
Derivative of minus log likelihood with respect to
one parameter: \code{y0}
}
\examples{
R code here showing how your function works
}
\keyword{keywords}
