% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_parametric_estimates.R
\name{non_parametric_estimates}
\alias{non_parametric_estimates}
\title{Run optimization on simulated data}
\usage{
non_parametric_estimates(survival_times, observed, continuous = FALSE)
}
\description{
Function doing the whole shebang.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}
