% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optimization_on_simulated_data.R
\name{run_optimization_on_simulated_data}
\alias{run_optimization_on_simulated_data}
\title{Run optimization on simulated data}
\usage{
run_optimization_on_simulated_data()
}
\description{
Function doing the whole shebang.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}
