% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optimization_on_test_data.R
\name{run_optimization_on_test_data}
\alias{run_optimization_on_test_data}
\title{Run optimization on test data}
\usage{
run_optimization_on_test_data()
}
\description{
Some description
}
\examples{
none
}
\keyword{keywords}
