% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_FHT_data.R
\name{simulate_FHT_data}
\alias{simulate_FHT_data}
\title{Simulate FHT data}
\usage{
simulate_FHT_data()
}
\value{
\code{list} A list of lots of things

\code{list} A list of lots of things
among others this and this
}
\description{
Simulate FHT data
}
\examples{
result <- generate_test_data()
observations <- result$observations
times <- observations$survival_times
delta <- observations$delta
# make
}
\keyword{keywords}
