% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optimization_on_test_data.R
\name{run_optimization_on_test_data}
\alias{run_optimization_on_test_data}
\title{Run optimization on test data}
\usage{
run_optimization_on_test_data()
}
\description{
Function doing the whole shebang.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}
