% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_FHT_data.R
\name{simulate_FHT_data}
\alias{simulate_FHT_data}
\title{Simulate FHT data}
\usage{
simulate_FHT_data(N = 1000, setup_type = "small_dense",
  add_noise = FALSE, seed = 2)
}
\value{
\code{observations} A list of \code{survival_times} and corresponding \code{delta}

\code{true_parameters} A list of \code{beta} and \code{gamma}

\code{design_matrices} A list of \code{X} and \code{Z}
}
\description{
Simulate FHT data
}
\examples{
result <- generate_test_data()
observations <- result$observations
times <- observations$survival_times
delta <- observations$delta
# make
}
\keyword{keywords}
